pyproject.toml
[project]
name = "yanscloud-service"
version = "0.1.1"
description = ""
authors = [
    {name = "imastarshine",email = "starshine@imastarshine.ru"}
]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv (>=1.2.1,<2.0.0)",
    "yadisk (>=3.4.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "soundcloudpy (>=0.1.4,<0.2.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "beautifulsoup4 (>=4.14.3,<5.0.0)",
    "aiohttp-socks (>=0.11.0,<0.12.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "curl-cffi (>=0.14.0,<0.15.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false

[tool.pyright]
venvPath = "."
venv = ".venv"
